### 2020-10-13 #2746 @whedon
[PRE REVIEW]: kalepy: a python package for kernel density estimation and sampling
**Submitting author:** @lzkelley (<a href="http://orcid.org/0000-0002-6625-6450">Luke Zoltan Kelley</a>)
**Repository:** <a href="https://github.com/lzkelley/kalepy" target ="_blank">https://github.com/lzkelley/kalepy</a>
**Version:** v1.0
