### 2020-10-14 #2746 @arfon
:wave: @lzkelley, thanks for your submission to JOSS! A few questions to get us started here:

- Is this submission associated with a [AAS publication](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing)?
- Could you please take a look at the DOI suggestions from Whedon [above](https://github.com/openjournals/joss-reviews/issues/2746#issuecomment-707890824) and if they are correct, add them to your BibTeX file?
- Do you have any suggestions for people who might be able to give this submission a good review?
