### 2020-10-14 #2746 @arfon
The software (_kalepy: a python package for kernel density estimation and sampling_) is [here](https://github.com/lzkelley/kalepy) and the short paper [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02746/joss.02746/10.21105.joss.02746.pdf).
