### 2020-10-15 b88b521
Add a couple of missing DOIs
