+++ paper/paper.bib
@@ -24,6 +24,7 @@
   number={3},
   pages={335--354},
   year={1977},
+  doi={10.1093/imamat/20.3.335},
   publisher={Oxford University Press}
 }
 
@@ -156,6 +157,7 @@
   for application development, interactive scripting, and
   publication-quality image generation across user
   interfaces and operating systems.},
+  doi={10.1109/mcse.2007.55},
   publisher = {IEEE COMPUTER SOC},
   year      = 2007
 }
