### 2020-10-15 #2746 @lzkelley
Hi @arfon sorry for the slow reply, and thanks for handling my submission!

> 👋 @lzkelley, thanks for your submission to JOSS! A few questions to get us started here:
>
> * Is this submission associated with a [AAS publication](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing)?
No, not in particular.
> * Could you please take a look at the DOI suggestions from Whedon [above](https://github.com/openjournals/joss-reviews/issues/2746#issuecomment-707890824) and if they are correct, add them to your BibTeX file?
Two of them were correct and have been added (in the `dev` branch).
> * Do you have any suggestions for people who might be able to give this submission a good review?
Unfortunately I don't.  I only know a few people who have worked on open-source projects and I don't think they'd be particularly well suited for this in particular.
