### 2020-11-12 89b22e0
BUG: convert input to array
