### 2020-11-12 #2784 @kexinrong
@lzkelley Nice package and visualizations! I will collect feedback on the repository here as I go through the review:

**Documentation**
- [ ] Could you compile the package documentation and add a link in the README?
- [ ] The README has a few incomplete sections.
- [ ] Since resampling is one of the main use cases, I will suggest exposing more parameters in the example usage. For example, for resample(), how do you specific size and weights? For KDE, what are the options for bandwidth.
- [ ] In the example "Plotting Smooth Distributions", the comment about 'points' is slightly confusing. Can you specify the granularity at which the PDF is evaluated?
- [ ] Missing guidelines for contribution and support.
