### 2020-11-16 #2784 @lzkelley
Thanks for the comments @kexinrong, these are all really helpful suggestions.  Could you clarify on your first point: "compile the package documentation" --- do you mean something in the 'readthedocs' style?
