### 2020-11-16 2ebe7e4
Cleanup requirements, separate into new 'dev-requirements.txt' file
