+++ requirements.txt
@@ -1,18 +1,4 @@
 numpy
 scipy
-
-codecov
 six
-future
-nose
-tqdm
-sphinx
-coverage
-memory_profiler
-nose-timer
-gitpython
-
-jupyter
-nbconvert
-corner
 matplotlib>=3
