### 2020-11-16 0abf227
BUG: 'weights' parameter was broken for resample without reflection
