### 2020-11-16 719c3b4
Add 'kalepy.plot.carpet' API function ('draw_carpet') still used internally
