### 2020-11-16 #2784 @kexinrong
@lzkelley Yes that's what I meant. I personally find that it is helpful to have readthedoc style API documentation, compared to inline documentation.
