### 2020-11-16 c60983f
Update travis for new dev-requirements file
