+++ docs/docs.sh
@@ -1,0 +1,4 @@
+# Generate documentation from in-package docstrings
+sphinx-apidoc -o ./source ../kalepy
+make clean
+make html
