+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+kalepy
+======
+
+.. toctree::
+   :maxdepth: 4
+
+   kalepy
