+++ docs/source/index.rst
@@ -1,55 +1,112 @@
-.. kdes documentation master file, created by
-   sphinx-quickstart on Mon May 20 14:31:16 2019.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+kalepy
+======
 
-Welcome to kdes's documentation!
-================================
+*Multidimensional kernel density estimation for calculating distribution functions and resampling.*
+
+Installation
+------------
+
+.. code-block:: bash
+
+    pip install kalepy
+
+
+or from source, for development:
+
+.. code-block:: bash
+
+    git clone https://github.com/lzkelley/kalepy.git
+    pip install -e kalepy
+
+
+
+Quickstart
+----------
+
+One dimensional kernel density estimation:
+******************************************
+
+.. code-block:: python
+
+   import kalepy as kale
+   import matplotlib.pyplot as plt
+   # For a one-dimensional data-set:
+   #   If the evaluation `points` aren't provided then `kalepy` automatically generates them
+   points, density = kale.density(data, points=None)
+
+   plt.plot(points, density, 'k-', lw=2.0, alpha=0.8, label='KDE')
+
+.. image:: https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_8_0.png
+    :alt: my-picture1
+
+
+One dimensional resampling:
+***************************
+
+.. code-block:: python
+
+    # Draw new samples from the KDE reconstructed PDF
+    samples = kale.resample(data)
+
+    # Plot new samples
+    plt.hist(samples, density=True, alpha=0.5, label='new samples', color='0.65', edgecolor='b')
+
+.. image:: https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_11_0.png
+
+
+Multi-dimensional kernel density estimation:
+********************************************
+
+.. code-block:: python
+
+    # Construct a KDE instance from data, shaped (N, 3) for `N` data points, and 3 dimensions
+    kde = kale.KDE(data)
+
+    # Build a corner plot using the `kalepy` plotting submodule
+    corner = kale.plot.Corner(kde, figsize=[10, 10])
+
+.. image:: https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_13_1.png
+
+
+
+Documentation
+-------------
+
+A number of examples are included in `the package notebooks <https://github.com/lzkelley/kalepy/tree/master/notebooks>`_, and the `readme file <https://github.com/lzkelley/kalepy/blob/master/README.md>`_.  Some background information and references are included in `the JOSS paper <>`_.
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
-
-
-Theory
-======
-
-Introduction
-------------
-
-A **Kernel Density Estimate (KDE)** is a generalization of a histogram, with two key differences.  First, the 'bins' are no longer restricted to boxes (which are zero within the bin-width of a data-point, and zero elsewhere), and are instead smoothing **'kernels'** that can take on an arbitrary functional form (they could still be a box, but Gaussian, and other shapes are also common).  Second, the location of the bins are no longer arbitrary and fixed, but instead are centered on the datapoints.
-
[…]
