+++ docs/source/kalepy.rst
@@ -1,0 +1,45 @@
+kalepy package
+==============
+
+Submodules
+----------
+
+kalepy.kde module
+-----------------
+
+.. automodule:: kalepy.kde
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+kalepy.kernels module
+---------------------
+
+.. automodule:: kalepy.kernels
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+kalepy.plot module
+------------------
+
+.. automodule:: kalepy.plot
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+kalepy.utils module
+-------------------
+
+.. automodule:: kalepy.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: kalepy
+   :members:
+   :undoc-members:
+   :show-inheritance:
