### 2020-12-09 d813440
Fix deprecations, docs updates
