+++ kalepy/utils.py
@@ -144,7 +144,7 @@ def check_path(fname):
 
 def cov_from_var_cor(var, corr):
     var = np.atleast_1d(var)
-    assert np.ndim(var) == 1, "`var` should be 1D!"
+    assert _ndim(var) == 1, "`var` should be 1D!"
     ndim = len(var)
     # Covariance matrix diagonals should be the variance (of each parameter)
     cov = np.identity(ndim) * var
@@ -183,7 +183,7 @@ def cumsum(vals, axis=None):
     """
 
     vals = np.asarray(vals)
-    nd = np.ndim(vals)
+    nd = _ndim(vals)
     if (axis is not None) or (nd == 1):
         return np.cumsum(vals, axis=axis)
 
@@ -226,7 +226,7 @@ def cumtrapz(pdf, edges, prepend=True, axis=None):
 
     # Prepend zeros to output array
     if prepend:
-        ndim = np.ndim(cdf)
+        ndim = _ndim(cdf)
         temp = [1, 0] if axis is None else [0, 0]
         padding = [temp for ii in range(ndim)]
         if axis is not None:
@@ -373,14 +373,14 @@ def parse_edges(data, edges=None, extrema=None, weights=None,
                 nmin=5, nmax=1000, pad=None, refine=1.0):
     """
     """
-    if np.ndim(data) not in [1, 2]:
+    if _ndim(data) not in [1, 2]:
         err = (
             "`data` (shape: {}) ".format(np.shape(data)) +
             "must have shape (N,) or (D, N) for `N` data points and `D` parameters!"
         )
         raise ValueError(err)
 
-    squeeze = (np.ndim(data) == 1)
+    squeeze = (_ndim(data) == 1)
     data = np.atleast_2d(data)
     npars = np.shape(data)[0]
 
@@ -390,7 +390,7 @@ def parse_edges(data, edges=None, extrema=None, weights=None,
     extrema = _parse_extrema(data, extrema=extrema, warn=False)
 
     # If `edges` provides a specification for each dimension, convert to npars*[edges]
-    if (np.ndim(edges) == 0) or (really1d(edges) and (np.size(edges) != npars)):
+    if (_ndim(edges) == 0) or (really1d(edges) and (np.size(edges) != npars)):
         edges = [edges] * npars
     elif len(edges) != npars:
         err = "length of `edges` ({}) does not match number of data dimensions ({})!".format(
@@ -424,7 +424,7 @@ def _get_edges_1d(edges, data, extrema, ndim, nmin, nmax, pad, weights=None, ref
 
     """
 
-    if np.ndim(edges) == 0:
+    if _ndim(edges) == 0:
         num_bins = edges
     elif really1d(edges):
         return edges
@@ -534,7 +534,7 @@ def quantiles(values, percs=None, sigmas=None, weights=None, axis=None, values_s
     if percs is None:
         percs = sp.stats.norm.cdf(sigmas)
 
-    if np.ndim(values) > 1:
+    if _ndim(values) > 1:
         if axis is None:
             values = values.flatten()
 
@@ -580,13 +580,18 @@ def really1d(arr):
         Whether `arr` is purely 1D.
 
     """
-    if np.ndim(arr) != 1:
+    if _ndim(arr) != 1:
         return False
     # Empty list or array
     if len(arr) == 0:
         return True
-    if np.any(np.vectorize(np.ndim)(arr)):
+    # if np.any(np.vectorize(_ndim)(arr)):
+    #     return False
+    # if len(arr) != flatlen(arr):
+    #     return False
+    if np.any([np.shape(tt) != () for tt in arr]):
[…]
