### 2020-12-24 #2784 @williamjameshandley
This is a very high-quality JOSS submission, with the code in an already very polished state, and a paper that is appropriate in content and length. The testing and CI are good, the installation smooth (both from PyPi and from source), and the examples in the README.md all work and clearly demonstrate the utility of the code. This is a package that I will likely incorporate into  [one of my own](https://github.com/williamjameshandley/anesthetic) as an optional dependency alternative to scipy KDE. Please could you raise kalepy integration as an issue on that repo to remind me.

I agree with @kexinrong that given all the work that has been put into the excellent numpy docstrings this should be continuously integrated readthedocs host using e.g. sphinx.

The paper itself has a couple of minor typos (Guassian -> Gaussian, Fig.Figure 1 -> Figure 1).

I think there are still a few missing README sections at the bottom, that need to be filled out. To fulfil community guidelines, I recommend going to the relevant [insights tab](https://github.com/lzkelley/kalepy/community) and following the examples there.

The only thing I think is missing from this is a discussion/comparison with what I would consider the present state-of-the-art in kernel density estimation in astronomy [GetDist](https://arxiv.org/pdf/1910.13970.pdf). kalepy includes resampling (which if possible in GetDist is not obvious to me how to do), but GetDist has a more sophisticated boundary correction system.

@arfon, I can't tick off my sections (I've been doing them locally for now). Could you re-send the invitation?
