### 2020-12-24 8c62a29
Improve random data routines
