### 2020-12-24 66f3a47
Restructure plotting module
