+++ kalepy/plot.py
@@ -1,12 +1,9 @@
 """
-
-BUG: `density` is being used differently in `corner_data` than in `dist1d_data` and `dist2d_data`!
-
 """
 
 # import warnings
 import logging
-import os
+# import os
 import six
 
 import numpy as np
@@ -29,80 +26,118 @@ _COLOR_CMAP = {
     'o': 'Oranges',
     'p': 'Purples',
 }
-_DEF_SIGMAS = np.arange(0.5, 2.1, 0.5)
+# _DEF_SIGMAS = np.arange(0.5, 2.1, 0.5)
+_DEF_SIGMAS = [0.5, 1.0, 1.5, 2.0]
 _MASK_CMAP = mpl.colors.LinearSegmentedColormap.from_list(
     "white_cmap", [(1, 1, 1), (1, 1, 1)], N=2)
 
-
-class plot_control:
-
-    def __init__(self, fname, *args, **kwargs):
-        self.fname = fname
-        self.args = args
-        self.kwargs = kwargs
-        return
-
-    def __enter__(self):
-        plt.close('all')
-        return self
-
-    def __exit__(self, type, value, traceback):
-        plt.savefig(self.fname, *self.args, **self.kwargs)
-        if utils._is_notebook():
-            plt.show()
-        else:
-            plt.close('all')
-
-        return
+_OUTLINE = ([
+    mpl.patheffects.Stroke(linewidth=2.0, foreground='white', alpha=0.85),
+    mpl.patheffects.Normal()
+])
 
 
 class Corner:
 
-    def __init__(self, kde_data, weights=None, labels=None, **figax_kwargs):
+    _LW = 2.0
+    _LIMITS_STRETCH = 0.1
+
+    def __init__(self, kde_data, weights=None, labels=None, limits=None, rotate=None, **kwfig):
+        """Construct a figure and axes for creating a 'corner' plot.
+
+        Arguments
+        ---------
+        kde_data : object, one of the following
+            * instance of `kalepy.kde.KDE`, including the data to be plotted,
+            * array_like scalar (D,N) of data with `D` parameters and `N` data points,
+            * int D, the number of parameters to construct a DxD corner plot.
+
+        weights : array_like scalar (N,) or None
+            The weights for each data point.
+            NOTE: only applicable when `kde_data` is a (D,N) dataset.
+
+        labels : array_like string (N,) of labels/names for each parameters.
+
+        limits : None or (2,) or (D,) of [None or (2,)]
+            Specification for the limits of each axes (for each of `D` parameters):
+            * None : the limits are determined automatically,
+            * (2,) : limits to be applied to all axes,
+            * (3,) : limits for each axis, where each entry can be either 'None' or (2,)
+
+        **kwfig : keyword-arguments passed to `_figax()` for constructing figure and axes.
+            See `kalepy.plot._figax()` for specifications.
+
+        """
+
+        # --- Parse the given `kde_data` and store parameters accordingly --
         kde = None
         data = None
+        # If a KDE instance is given, take all parameters from there
         if isinstance(kde_data, kale.KDE):
+            if weights is not None:
+                raise ValueError("`weights` must be used from given `KDE` instance!")
             kde = kde_data
             data = kde.dataset
             weights = None if kde._weights_uniform else kde.weights
             size = len(data)
+        # If an integer is given, it's the number of parameters/variables
         elif np.isscalar(kde_data):
+            if not isinstance(kde_data, int):
+                err = ("If `kde_data` is a scalar, it must be an integer "
+                       "specifying the number of parameters!")
+                raise ValueError(err)
             size = kde_data
[…]
