### 2020-12-26 #2784 @arfon
> @arfon, I can't tick off my sections (I've been doing them locally for now). Could you re-send the invitation?

@williamjameshandley - you should be good to go once you've accepted the invite at the link above.
