### 2020-12-26 b28b09d
Allow 'params=0' to be passed for 1D data
