### 2020-12-26 1bde665
Plotting restructuring
