+++ kalepy/plot.py
@@ -4,7 +4,7 @@
 # import warnings
 import logging
 # import os
-import six
+# import six
 
 import numpy as np
 import scipy as sp
@@ -17,20 +17,9 @@ import kalepy as kale
 from kalepy import utils
 from kalepy import KDE
 
-_STRETCH = 0.2
-_COLOR_CMAP = {
-    'k': 'Greys',
-    'b': 'Blues',
-    'r': 'Reds',
-    'g': 'Greens',
-    'o': 'Oranges',
-    'p': 'Purples',
-}
-# _DEF_SIGMAS = np.arange(0.5, 2.1, 0.5)
 _DEF_SIGMAS = [0.5, 1.0, 1.5, 2.0]
 _MASK_CMAP = mpl.colors.LinearSegmentedColormap.from_list(
     "white_cmap", [(1, 1, 1), (1, 1, 1)], N=2)
-
 _OUTLINE = ([
     mpl.patheffects.Stroke(linewidth=2.0, foreground='white', alpha=0.85),
     mpl.patheffects.Normal()
@@ -70,40 +59,25 @@ class Corner:
         """
 
         # --- Parse the given `kde_data` and store parameters accordingly --
-        kde = None
-        data = None
-        # If a KDE instance is given, take all parameters from there
-        if isinstance(kde_data, kale.KDE):
-            if weights is not None:
-                raise ValueError("`weights` must be used from given `KDE` instance!")
-            kde = kde_data
-            data = kde.dataset
-            weights = None if kde._weights_uniform else kde.weights
-            size = len(data)
-        # If an integer is given, it's the number of parameters/variables
-        elif np.isscalar(kde_data):
+        if np.isscalar(kde_data):
             if not isinstance(kde_data, int):
                 err = ("If `kde_data` is a scalar, it must be an integer "
                        "specifying the number of parameters!")
                 raise ValueError(err)
             size = kde_data
-        # If the raw data is given, construct a KDE from it
+            kde = None
+            data = None
         else:
-            try:
-                kde = KDE(kde_data, weights=weights)
-            except:
-                err = "Failed to construct `KDE` instance from given data!"
-                logging.error(err)
-                raise
-
-            data = kde_data
-            size = len(data)
+            kde, data, weights = _parse_kde_data(kde_data, weights=weights)
+            size = kde._ndim
 
         if limits is None:
             limits = [None] * size
+            limit_flag = True
+        else:
+            limit_flag = False
 
         if rotate is None:
-            # rotate = (size == 2)
             rotate = (size in [2, 3])
 
         # -- Construct figure and axes using `_figax()`
@@ -155,66 +129,261 @@ class Corner:
 
         # --- Store key parameters
         self.size = size
+        self._kde = kde
         self._data = data
         self._weights = weights
-        self._kde = kde
         self._limits = limits
+        self._limit_flag = limit_flag
         self._rotate = rotate
 
         return
 
-    def plot(self, kde=None, data=None, kwkde=None, kwdata=None, **kwargs):
+    def plot(self, kde_data=None, edges=None, weights=None, rotate=None, quantiles=None,
+             limit=None, color=None, cmap=None, dist1d={}, dist2d={}):
+
+        if kde_data is None:
+            # If either data or a KDE was given on initialization, `self._kde` will be set
+            kde_data = self._kde
+            if kde_data is None:
+                err = "kde or data required either during initialization or here!"
+                raise ValueError(err)
+
+        kde, data, weights = _parse_kde_data(kde_data, weights=weights)
+
+        dist1d.setdefault('density', True)
[…]
