### 2020-12-26 f0b3113
BUG: fix issue with axes limits not being set properly; restrict colormap range for contours; cleanup
