+++ kalepy/plot.py
@@ -18,17 +18,10 @@ from kalepy import utils
 from kalepy import KDE
 
 _DEF_SIGMAS = [0.5, 1.0, 1.5, 2.0]
-_MASK_CMAP = mpl.colors.LinearSegmentedColormap.from_list(
-    "white_cmap", [(1, 1, 1), (1, 1, 1)], N=2)
-_OUTLINE = ([
-    mpl.patheffects.Stroke(linewidth=2.0, foreground='white', alpha=0.85),
-    mpl.patheffects.Normal()
-])
 
 
 class Corner:
 
-    _LW = 2.0
     _LIMITS_STRETCH = 0.1
 
     def __init__(self, kde_data, weights=None, labels=None, limits=None, rotate=None, **kwfig):
@@ -70,12 +63,6 @@ class Corner:
         else:
             kde, data, weights = _parse_kde_data(kde_data, weights=weights)
             size = kde._ndim
-
-        if limits is None:
-            limits = [None] * size
-            limit_flag = True
-        else:
-            limit_flag = False
 
         if rotate is None:
             rotate = (size in [2, 3])
@@ -127,6 +114,13 @@ class Corner:
             else:
                 pass
 
+        if limits is None:
+            limits = [None] * size
+            limit_flag = True
+        else:
+            limit_flag = False
+            _set_corner_axes_extrema(self.axes, limits, rotate)
+
         # --- Store key parameters
         self.size = size
         self._kde = kde
@@ -168,7 +162,7 @@ class Corner:
         )
         return rv
 
-    def clean(self, kde_data=None, weights=None, **kwargs):
+    def clean(self, kde_data=None, weights=None, dist1d={}, dist2d={}, **kwargs):
         if kde_data is None:
             # If either data or a KDE was given on initialization, `self._kde` will be set
             kde_data = self._kde
@@ -424,229 +418,8 @@ class Corner:
     '''
 
 
-def _figax(size, grid=True, left=None, bottom=None, right=None, top=None, hspace=None, wspace=None,
-           **kwfig):
-
-    _def_figsize = np.clip(4 * size, 6, 20)
-    _def_figsize = [_def_figsize for ii in range(2)]
-
-    figsize = kwfig.pop('figsize', _def_figsize)
-    if not np.iterable(figsize):
-        figsize = [figsize, figsize]
-
-    if hspace is None:
-        hspace = 0.1
-    if wspace is None:
-        wspace = 0.1
-
-    fig, axes = plt.subplots(figsize=figsize, squeeze=False, ncols=size, nrows=size, **kwfig)
-
-    plt.subplots_adjust(
-        left=left, bottom=bottom, right=right, top=top, hspace=hspace, wspace=wspace)
-    if grid is True:
-        grid = dict(alpha=0.2, color='0.5', lw=0.5)
-    elif grid is False:
-        grid = None
-
-    for idx, ax in np.ndenumerate(axes):
-        if grid is not None:
-            ax.grid(True, **grid)
-
-    return fig, axes
-
-
-def corner(kde_data, labels=None, kwcorner={}, kwplot={}):
-    corner = Corner(kde_data, labels=labels, **kwcorner)
-    corner.plot(**kwplot)
-    return corner
-
-
 # ======  API Methods  ======
 # ================================
-
-
-def dist2d(kde_data, ax=None, edges=None, weights=None,
-           params=[0, 1], probability=True, quantiles=None, color=None, cmap=None,
-           median=True, scatter=True, contour=True, hist=True, mask_dense=None, mask_below=True):
-
-    if ax is None:
-        ax = plt.gca()
-
-    if isinstance(kde_data, kale.KDE):
-        if weights is not None:
[…]
