### 2020-12-26 051b6f2
Plotting cleanup and updates
