+++ kalepy/plot.py
@@ -1,14 +1,9 @@
 """
 """
 
-# import warnings
 import logging
-# import os
-# import six
 
 import numpy as np
-import scipy as sp
-import scipy.stats  # noqa
 import matplotlib as mpl
 import matplotlib.patheffects  # noqa
 import matplotlib.pyplot as plt
@@ -18,6 +13,13 @@ from kalepy import utils
 from kalepy import KDE
 
 _DEF_SIGMAS = [0.5, 1.0, 1.5, 2.0]
+_PAD = 1
+
+# Define functions for module level import
+__all__ = [
+    "carpet", "contour", "confidence", "Corner", "corner", "dist1d", "dist2d", "hist1d", "hist2d",
+    "plot_clean", "plot_data", "plot_hist", "plot_kde",
+]
 
 
 class Corner:
@@ -145,7 +147,7 @@ class Corner:
         kde, data, weights = _parse_kde_data(kde_data, weights=weights)
 
         dist1d.setdefault('density', True)
-        dist1d.setdefault('contour', True)
+        dist1d.setdefault('confidence', True)
         dist1d.setdefault('carpet', True)
         dist1d.setdefault('hist', False)
 
@@ -173,7 +175,7 @@ class Corner:
         kde, data, weights = _parse_kde_data(kde_data, weights=weights)
 
         dist1d.setdefault('density', True)
-        dist1d.setdefault('contour', False)
+        dist1d.setdefault('confidence', False)
         dist1d.setdefault('carpet', False)
         dist1d.setdefault('hist', False)
 
@@ -196,7 +198,7 @@ class Corner:
         kde, data, weights = _parse_kde_data(kde_data, weights=weights)
 
         dist1d.setdefault('density', False)
-        dist1d.setdefault('contour', False)
+        dist1d.setdefault('confidence', False)
         dist1d.setdefault('carpet', False)
         dist1d.setdefault('hist', True)
 
@@ -279,7 +281,7 @@ class Corner:
 
     def _data1d(self, ax, edge, data, color=None, **dist1d):
         dist1d.setdefault('density', False)
-        dist1d.setdefault('contour', False)
+        dist1d.setdefault('confidence', False)
         dist1d.setdefault('carpet', True)
         dist1d.setdefault('hist', True)
         rv = _dist1d(data, ax=ax, edges=edge, color=color, **dist1d)
@@ -361,7 +363,7 @@ class Corner:
 
     def _kde1d(self, ax, edge, kde, param, color=None, **dist1d):
         dist1d.setdefault('density', True)
-        dist1d.setdefault('contour', True)
+        dist1d.setdefault('confidence', True)
         dist1d.setdefault('carpet', False)
         dist1d.setdefault('hist', False)
         # This is identical to `kalepy.plot.dist1d` (just used for naming convenience)
@@ -418,8 +420,38 @@ class Corner:
     '''
 
 
-# ======  API Methods  ======
-# ================================
+def corner(kde_data, labels=None, kwcorner={}, **kwplot):
+    corner = Corner(kde_data, labels=labels, **kwcorner)
+    corner.plot(**kwplot)
+    return corner
+
+
+def plot_clean(kde_data, labels=None, kwcorner={}, **kwplot):
+    corner = Corner(kde_data, labels=labels, **kwcorner)
+    corner.clean(**kwplot)
+    return corner
+
+
+def plot_data(data, labels=None, kwcorner={}, **kwplot):
+    corner = Corner(data, labels=labels, **kwcorner)
+    corner.plot_data(**kwplot)
+    return corner
+
+
+def plot_hist(kde_data, labels=None, kwcorner={}, **kwplot):
[…]
