+++ kalepy/__init__.py
@@ -26,11 +26,11 @@ _BANDWIDTH_DEFAULT = 'scott'
 _PATH_NB = os.path.join(_path, os.path.pardir, 'notebooks', '')
 _PATH_NB_OUT = os.path.join(_PATH_NB, 'output', '')
 
-from kalepy import kernels   # noqa
-from kalepy import utils   # noqa
+from kalepy import kernels  # noqa
+from kalepy import utils    # noqa
 from kalepy.kde import KDE  # noqa
-import kalepy.plot  # noqa
-from kalepy.plot import corner, Corner  # noqa
+import kalepy.plot          # noqa
+from kalepy.plot import *   # noqa
 
 del os
 del inn
