### 2020-12-27 0c9ccd8
Allow some random data to be generalized
