### 2020-12-27 71effc3
Add extensive plotting docstrings
