+++ kalepy/plot.py
@@ -1,4 +1,32 @@
-"""
+"""kalepy's plotting submodule
+
+This submodule containts the `Corner` class, and all plotting methods.  The `Corner` class, and
+additional API functions are imported into the base package namespace of `kalepy`, e.g.
+`kalepy.Corner` and `kalepy.carpet` access the `kalepy.plot.Corner` and `kalepy.plot.carpet` methods
+respectively.
+
+Additional options and customization:
+
+The core plotting routines, such as `draw_hist1d`, `draw_hist2d`, `draw_contour2d`, etc include a
+fairly large number of keyword arguments for customization.  The top level API methods, such as
+`corner()` or `Corner.plot_data()` often do not provide access to all of those arguments, but
+additional customization is possible by using the drawing methods directly, and optionally
+subclassing the `Corner` class to provide additional or different functionality.
+
+API Contents
+------------
+- Corner : class for corner/triangle/pair plots.
+- corner : method which constructs a `Corner` instance and plots 1D and 2D distributions.
+
+- dist1d : plot a 1D distribution with numerous possible elements (e.g. histogram, carpet, etc)
+- dist2d : plot a 2D distribution with numerous possible elements (e.g. histogram, contours, etc)
+
+- carpet : draw a 1D scatter-like plot to depict semi-quantitative information about a distribution.
+- contour : draw a 2D contour plot. A wrapper of additional functionality around `plt.contour`
+- confidence : draw 1D confidence intervals using shaded bands.
+- hist1d : draw a 1D histogram
+- hist2d : draw a 2D histogram.  A wrapper of additional functionality around `plt.pcolormesh`
+
 """
 
 import logging
@@ -18,35 +46,72 @@ _PAD = 1
 # Define functions for module level import
 __all__ = [
     "carpet", "contour", "confidence", "Corner", "corner", "dist1d", "dist2d", "hist1d", "hist2d",
-    "plot_clean", "plot_data", "plot_hist", "plot_kde",
 ]
 
 
 class Corner:
+    """Class for creating 'corner' / 'pair' plots of multidimensional covariances.
 
+    The `Corner` class acts as a constructor for a `matplotlib` figure and axes, and coordinates the
+    plotting of 1D and 2D distributions.  The `kalepy.plot.dist1d()` and `kalepy.plot.dist2d()`
+    methods are used for plotting the distributions.  The class methods provide wrappers, and
+    default setting for those methods.  The `Corner.plot` method is the standard plotting method
+    with default parameters chosen for plotting a single, multidimensional dataset.  For
+    overplotting numerous datasets, the `Corner.clean` or `Corner.plot_data` methods are better.
+
+    API Methods
+    -----------
+    - `plot` : the standard plotting method which, by default, includes both KDE and data elements.
+    - `clean` : minimal plots with only the KDE generated PDF in 1D and contours in 2D, by default.
+    - `hist` : minimal plots with only the data based 1D and 2D histograms, by default.
+    - `plot_kde` : plot elements with only KDE based info: the `clean` settings with a little more.
+    - `plot_data` : plot elements without using KDE info.
+
+    Examples
+    --------
+    Load some predefined 3D data, and generate a default corner plot:
+
+    >>> import kalepy as kale
+    >>> data = kale.utils._random_data_3d_03()
+    >>> kale.corner(data)
+
+    Load two different datasets, and overplot them using a `kalepy.Corner` instance.
+
+    >>> data1 = kale.utils._random_data_3d_03(par=[0.0, 0.5], cov=0.05)
+    >>> data2 = kale.utils._random_data_3d_03(par=[1.0, 0.25], cov=0.5)
[…]
