### 2020-12-27 252d3a6
DOCS: more docstrings in plotting
