+++ kalepy/plot.py
@@ -169,6 +169,7 @@ class Corner:
 
             # Diagonals
             if ii == jj:
+                # not top-left
                 if ii != 0:
                     ax.yaxis.set_label_position('right')
                     ax.yaxis.set_ticks_position('right')
@@ -177,12 +178,14 @@ class Corner:
             else:
                 pass
 
-        if limits is None:
-            limits = [None] * size
-            limit_flag = True
-        else:
+        # If axes limits are given, set axes to them
+        if limits is not None:
             limit_flag = False
             _set_corner_axes_extrema(self.axes, limits, rotate)
+        # Otherwise, prepare to calculate limits during plotting
+        else:
+            limits = [None] * size
+            limit_flag = True
 
         # --- Store key parameters
         self.size = size
@@ -223,20 +226,25 @@ class Corner:
                 err = "kde or data required either during initialization or here!"
                 raise ValueError(err)
 
+        # If data is given, convert to KDE as needed
         kde, data, weights = _parse_kde_data(kde_data, weights=weights)
 
+        # Set default 1D parameters
         dist1d.setdefault('density', True)
         dist1d.setdefault('confidence', False)
         dist1d.setdefault('carpet', False)
         dist1d.setdefault('hist', False)
 
+        # Set default 2D parameters
         dist2d.setdefault('hist', False)
         dist2d.setdefault('contour', True)
         dist2d.setdefault('scatter', False)
         dist2d.setdefault('mask_dense', False)
         dist2d.setdefault('mask_below', False)
 
-        return self.plot_kde(kde, dist1d=dist1d, dist2d=dist2d, **kwargs)
+        # Plot
+        rv = self.plot_kde(kde, dist1d=dist1d, dist2d=dist2d, **kwargs)
+        return rv
 
     def hist(self, kde_data=None, weights=None, dist1d={}, dist2d={}, **kwargs):
         """Wrapper for `plot_data` that sets parameters to only plot 1D and 2D histograms of data.
@@ -266,20 +274,25 @@ class Corner:
                 err = "kde or data required either during initialization or here!"
                 raise ValueError(err)
 
+        # If KDE is given, retrieve the dataset from it
         kde, data, weights = _parse_kde_data(kde_data, weights=weights)
 
+        # Set default 1D parameters
         dist1d.setdefault('density', False)
         dist1d.setdefault('confidence', False)
         dist1d.setdefault('carpet', False)
         dist1d.setdefault('hist', True)
 
+        # Set default 2D parameters
         dist2d.setdefault('hist', True)
         dist2d.setdefault('contour', False)
         dist2d.setdefault('scatter', False)
         dist2d.setdefault('mask_dense', False)
         dist2d.setdefault('mask_below', False)
 
-        return self.plot_data(data, dist1d=dist1d, dist2d=dist2d, **kwargs)
+        # Plot
+        rv = self.plot_data(data, dist1d=dist1d, dist2d=dist2d, **kwargs)
+        return rv
 
     def plot(self, kde_data=None, edges=None, weights=None, quantiles=None,
              limit=None, color=None, cmap=None, dist1d={}, dist2d={}):
@@ -343,19 +356,23 @@ class Corner:
                 err = "kde or data required either during initialization or here!"
                 raise ValueError(err)
 
+        # If data is given, construct KDE
         kde, data, weights = _parse_kde_data(kde_data, weights=weights)
 
+        # Set default 1D parameters
         dist1d.setdefault('density', True)
[…]
