+++ .travis.yml
@@ -1,10 +1,10 @@
 language: python
 python:
-  - "3.5"
-  - "3.5-dev"
   - "3.6"
   - "3.6-dev"
   - "3.7-dev"
+  - "3.8"
+  - "3.8-dev"
 
 # Enable 3.7 without globally enabling sudo and dist: xenial for other build jobs
 # See: https://github.com/travis-ci/travis-ci/issues/9815
