### 2020-12-28 70d93ed
Cleanup; avoid numpy deprecation warning; dont use f-strings for older python compat
