### 2020-12-28 db5684b
DOCS: lots of documentation updates
