+++ kalepy/plot.py
@@ -207,7 +207,7 @@ class Corner:
             * instance of `kalepy.kde.KDE`, providing the data and KDE to be plotted.
             * array_like scalar (D,N) of data with `D` parameters and `N` data points.
             * `None` : use the KDE/data stored during class initialization.
-                       raises `ValueError` if no KDE/data was provided
+              raises `ValueError` if no KDE/data was provided
 
         weights : `None` or (N,) array_like of scalar, the weighting of each data-point if and
             only-if the given `kde_data` argument is a (D,N) array_like of scalar data.
@@ -255,7 +255,7 @@ class Corner:
             * instance of `kalepy.kde.KDE`, providing the data and KDE to be plotted.
             * array_like scalar (D,N) of data with `D` parameters and `N` data points.
             * `None` : use the KDE/data stored during class initialization.
-                       raises `ValueError` if no KDE/data was provided
+              raises `ValueError` if no KDE/data was provided
 
         weights : `None` or (N,) array_like of scalar, the weighting of each data-point if and
             only-if the given `kde_data` argument is a (D,N) array_like of scalar data.
@@ -309,14 +309,14 @@ class Corner:
             * instance of `kalepy.kde.KDE`, providing the data and KDE to be plotted.
             * array_like scalar (D,N) of data with `D` parameters and `N` data points.
             * `None` : use the KDE/data stored during class initialization.
-                       raises `ValueError` if no KDE/data was provided
+              raises `ValueError` if no KDE/data was provided
 
         edges : object specifying historgam edge locations; or `None`
             * int : the number of bins for all dimensions, locations calculated automatically
             * (D,) array_like of int : the number of bins for each of `D` dimensions
             * (D,) of array_like : the bin-edge locations for each of `D` dimensions, e.g.
-                                   ([0, 1, 2], [0.0, 0.1, 0.2, 0.3],) would describe two bins for
-                                   the 0th dimension, and 3 bins for the 1st dimension.
+              ([0, 1, 2], [0.0, 0.1, 0.2, 0.3],) would describe two bins for
+              the 0th dimension, and 3 bins for the 1st dimension.
             * (X,) array_like of scalar : the bin-edge locations to be used for all dimensions
             * `None` : the number and locations of bins are calculated automatically for each dim
 
@@ -334,13 +334,14 @@ class Corner:
             If `None`:
                 * `cmap` is given, then the color will be set to the `cmap` midpoint.
                 * `cmap` is not given, then the color will be determined by the next value of the
-                    default matplotlib color-cycle, and `cmap` will be set to a matching colormap.
+                  default matplotlib color-cycle, and `cmap` will be set to a matching colormap.
+
             This parameter effects the color of 1D: histograms, confidence intervals, and carpet;
             2D: scatter points.
 
         cmap : matplotlib colormap specification, or `None`
             * All valid matplotlib specifications can be used, e.g. named value (like 'Reds' or
-            'viridis') or a `matplotlib.colors.Colormap` instance.
+              'viridis') or a `matplotlib.colors.Colormap` instance.
             * If `None` then a colormap is constructed based on the value of `color` (see above).
 
         dist1d : dict of keyword-arguments passed to the `kale.plot.dist1d` method.
@@ -394,14 +395,14 @@ class Corner:
             * instance of `kalepy.kde.KDE`, providing the data and KDE to be plotted.
             * array_like scalar (D,N) of data with `D` parameters and `N` data points.
             * `None` : use the KDE/data stored during class initialization.
[…]
