+++ docs/source/conf.py
@@ -31,10 +31,11 @@ release = '0.4'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'nbsphinx',
+    'nbsphinx',                # convert notebooks to sphinx output
     'sphinx.ext.napoleon',     # allow numpy/google style docstrings
     'sphinx.ext.autodoc',      # auto-generate documentation from docstrings
-    'sphinx.ext.mathjax'       # render math in html using mathjax
+    'sphinx.ext.mathjax',      # render math in html using mathjax
+    'sphinx.ext.autosummary'
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -62,11 +63,18 @@ html_static_path = ['_static']
 
 # -- Extensions Parameters ----------------------------------------------------
 
+# autodoc
+autoclass_content = 'both'
+autodoc_default_options = {
+    'autoclass_content': 'both',
+    'special-members': '__init__',
+}
+
 # Napoleon settings
 napoleon_google_docstring = True
 napoleon_numpy_docstring = True
 napoleon_include_init_with_doc = False
-napoleon_include_private_with_doc = True
+napoleon_include_private_with_doc = False
 napoleon_include_special_with_doc = True
 napoleon_use_admonition_for_examples = False
 napoleon_use_admonition_for_notes = False
