+++ README.md
@@ -2,7 +2,7 @@
 
 [![Build Status](https://travis-ci.org/lzkelley/kalepy.svg?branch=master)](https://travis-ci.org/lzkelley/kalepy)
 [![codecov](https://codecov.io/gh/lzkelley/kalepy/branch/master/graph/badge.svg)](https://codecov.io/gh/lzkelley/kalepy)
-
+[![Documentation Status](https://readthedocs.org/projects/kalepy/badge/?version=latest)](https://kalepy.readthedocs.io/en/latest/?badge=latest)
 
 ![kalepy animated logo](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/logo_anim_small.gif)
 
@@ -34,7 +34,7 @@ nosetests
 ```
 
 
-## Basic Usage
+# Basic Usage
 
 
 ```python
@@ -53,10 +53,12 @@ Generate some random data, and its corresponding distribution function
 ```python
 NUM = int(1e4)
 np.random.seed(12345)
+# Combine data from two different PDFs
 _d1 = np.random.normal(4.0, 1.0, NUM)
 _d2 = np.random.lognormal(0, 0.5, size=NUM)
 data = np.concatenate([_d1, _d2])
 
+# Calculate the "true" distribution
 xx = np.linspace(0.0, 7.0, 100)[1:]
 yy = 0.5*np.exp(-(xx - 4.0)**2/2) / np.sqrt(2*np.pi)
 yy += 0.5 * np.exp(-np.log(xx)**2/(2*0.5**2)) / (0.5*xx*np.sqrt(2*np.pi))
@@ -84,7 +86,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_8_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_10_0.png)
 
 
 ### Constructing statistically similar values
@@ -99,7 +101,6 @@ samples = kale.resample(data)
 # Plot new samples
 plt.hist(samples, density=True, alpha=0.5, label='new samples', color='0.65', edgecolor='b')
 
-
 # Plot the KDE reconstructed PDF
 plt.plot(points, density, 'k-', lw=2.0, alpha=0.8, label='KDE')
 # Plot the "true" PDF
@@ -110,7 +111,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_11_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_13_0.png)
 
 
 ### Multivariate Distributions
@@ -118,35 +119,33 @@ nbshow()
 
 ```python
 # Load some random-ish data
-data = kale.utils._random_data_3d_01()
+# data = kale.utils._random_data_3d_01()
+data = kale.utils._random_data_3d_02()
 
 # Construct a KDE
 kde = kale.KDE(data)
 
-import kalepy.plot
+# import kalepy.plot
 
 # Build a corner plot using the `kalepy` plotting submodule
 corner = kale.plot.Corner(kde, figsize=[10, 10])
 
 # Data points: red scatter and histograms
-corner.plot_data(color='red', scatter=dict(s=10, alpha=0.15))
+corner.plot_data(color='red')
 
 # KDE reconstructed density-distribution: blue contours and curves
 corner.plot_kde(color='blue')
 
-
-plt.show()
+nbshow()
 ```
 
[…]
