### 2020-12-28 007a6ae
BUG: Handle 'extrema' when only some 'params' are specified
