+++ gen_kde_api.py
@@ -1,0 +1,208 @@
+"""Generate the `plot_api.rst` file.
+"""
+
+import os
+import sys
+import shutil
+import re
+import git
+import subprocess
+import logging
+
+
+NOTEBOOK_FNAME = "./notebooks/demo.ipynb"
+NB_SUFFIX = "ipynb"
+# Path for output resources (i.e. image files)
+OUTPUT_PATH_RESOURCES = "./docs/media/"
+# Path for output markdown file
+OUTPUT_FNAME = "./docs/source/kde_api.rst"
+
+# Only the output of the notebook *after* this line, is written to output
+NB_START_HEADER = "demo"
+
+# CONVERT_IMAGE_REGEX = r"^\!\[png\]\((.*)\)"    # MARKDOWN
+CONVERT_IMAGE_REGEX = r"^\.\. image\:\: (.*)"   # RST
+GITHUB_RAW_ADDR = "https://raw.githubusercontent.com/lzkelley/kalepy/{branch:}/docs/media/"
+
+ALLOWED_BRANCHES = ['dev', 'master']
+
+EXECUTE = True
+
+
+def main():
+    global EXECUTE
+
+    if '-v' in sys.argv:
+        idx = sys.argv.index('-v')
+        level = 20
+        if len(sys.argv) > idx+1:
+            _lev = sys.argv[idx+1]
+            if not _lev.startswith('-'):
+                level = int(_lev)
+
+        logging.getLogger().setLevel(level)
+
+    if '-x' in sys.argv:
+        EXECUTE = True
+    elif '-nx' in sys.argv:
+        EXECUTE = False
+
+    logging.debug("\t`EXECUTE` = {}".format(EXECUTE))
+    logging.debug("sys.argv = '{}'".format(sys.argv))
+
+    # Load repository information
+    repo = git.Repo('.')
+    if repo.bare:
+        raise RuntimeError("Failed to initialize `git.Repo`!")
+
+    try:
+        branch = [bb for bb in repo.git.branch().split('\n') if bb.strip().startswith('* ')]
+        if len(branch) != 1:
+            raise
+        branch = branch[0].strip('*').strip()
+    except:
+        logging.error("Failed to get current git branch!")
+        raise
+
+    logging.info("Current branch: '{}'".format(branch))
+    if branch not in ALLOWED_BRANCHES:
+        raise ValueError("Current branch '{}' not expected!".format(branch))
+
+    # Check that input notebook exists
+    nb_fname = os.path.abspath(NOTEBOOK_FNAME)
+    exists = os.path.isfile(nb_fname)
+    logging.info("Input notebook: '{}' exists: {}".format(NOTEBOOK_FNAME, exists))
+    if not exists:
+        raise ValueError("Notebook '{}' does not exist!".format(nb_fname))
+
+    nb_base = nb_fname.split("." + NB_SUFFIX)
+    if len(nb_base) != 2:
+        raise ValueError("Could not split filename on '{}'".format(NB_SUFFIX))
+
+    out_fil_temp = nb_base[0] + ".rst"
+    out_dir_temp = nb_base[0] + "_files"
+    out_fil = os.path.abspath(OUTPUT_FNAME)
+    out_dir = os.path.basename(out_dir_temp)
+    # if `out_dir_temp` ends with a '/' then basename is empty string
+    if len(out_dir) == 0:
+        raise ValueError("Something wrong with output path!")
+    out_dir = os.path.join(os.path.abspath(OUTPUT_PATH_RESOURCES), out_dir, '')
+    if not os.path.exists(out_dir):
+        os.mkdir(out_dir)
+
+    # Remove previously added image files (in `out_dir`) from the git repo before deleting
+    for fil in os.listdir(out_dir):
[…]
