+++ gen_plot_api.py
@@ -24,7 +24,7 @@ NB_START_HEADER = "demo_plot"
 CONVERT_IMAGE_REGEX = r"^\.\. image\:\: (.*)"   # RST
 GITHUB_RAW_ADDR = "https://raw.githubusercontent.com/lzkelley/kalepy/{branch:}/docs/media/"
 
-ALLOWED_BRANCHES = ['dev', 'master']
+ALLOWED_BRANCHES = ['dev', 'master', 'test']
 
 EXECUTE = False
 
