### 2020-12-29 a223891
Re-enable 'Corner.legend()', pass along hist1d plot handle to/from Corner.plot methods
