### 2020-12-29 cc0517a
New random 1D data set
