### 2020-12-29 99835f3
Update 'demo.ipynb' and associates docs
