+++ README.md
@@ -69,7 +69,6 @@ yy += 0.5 * np.exp(-np.log(xx)**2/(2*0.5**2)) / (0.5*xx*np.sqrt(2*np.pi))
 
 ```python
 # Reconstruct the probability-density based on the given data points.
-# If `points` aren't provided then `kalepy` automatically generates them
 points, density = kale.density(data, probability=True)
 
 # Plot the PDF
@@ -89,7 +88,7 @@ nbshow()
 ![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_8_0.png)
 
 
-### Constructing statistically similar values
+### resampling: constructing statistically similar values
 
 Draw a new sample of data-points from the KDE PDF
 
@@ -99,12 +98,12 @@ Draw a new sample of data-points from the KDE PDF
 samples = kale.resample(data)
 
 # Plot new samples
-plt.hist(samples, density=True, alpha=0.5, label='new samples', color='0.65', edgecolor='b')
+plt.hist(samples, density=True, label='new samples', alpha=0.5, color='0.65', edgecolor='b')
+# Plot the old samples
+plt.hist(data, density=True, histtype='step', lw=2.0, alpha=0.5, color='r', label='input data')
 
 # Plot the KDE reconstructed PDF
 plt.plot(points, density, 'k-', lw=2.0, alpha=0.8, label='KDE')
-# Plot the "true" PDF
-plt.plot(xx, yy, 'r--', alpha=0.4, lw=3.0, label='truth')
 
 plt.legend()
 nbshow()
@@ -118,23 +117,14 @@ nbshow()
 
 
 ```python
-# Load some random-ish data
-# data = kale.utils._random_data_3d_01()
+# Load some random-ish three-dimensional data
 data = kale.utils._random_data_3d_02()
 
 # Construct a KDE
 kde = kale.KDE(data)
 
-# import kalepy.plot
-
-# Build a corner plot using the `kalepy` plotting submodule
-corner = kale.plot.Corner(kde, figsize=[10, 10])
-
-# Data points: red scatter and histograms
-corner.plot_data(color='red')
-
-# KDE reconstructed density-distribution: blue contours and curves
-corner.plot_kde(color='blue')
+# Plot the data and distributions using the builtin `kalepy.corner` plot
+kale.corner(kde)
 
 nbshow()
 ```
@@ -143,9 +133,36 @@ nbshow()
 ![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_13_0.png)
 
 
-# Fancy Usage
 
-## Distribution Functions
+```python
+# Resample the data (default output is the same size as the input data)
+samples = kde.resample()
+
+
+# ---- Plot the input data compared to the resampled data ----
+
+fig, axes = plt.subplots(figsize=[16, 4], ncols=kde.ndim)
+
+for ii, ax in enumerate(axes):
+    # Calculate and plot PDF for `ii`th parameter (i.e. data dimension `ii`)
+    xx, yy = kde.density(params=ii, probability=True)
+    ax.plot(xx, yy, 'k--', label='KDE', lw=2.0, alpha=0.5)
+    # Draw histograms of original and newly resampled datasets
+    *_, h1 = ax.hist(data[ii], histtype='step', density=True, lw=2.0, label='input')
+    *_, h2 = ax.hist(samples[ii], histtype='step', density=True, lw=2.0, label='resample')
+    # Add 'kalepy.carpet' plots showing the data points themselves
+    kale.carpet(data[ii], ax=ax, color=h1[0].get_facecolor())
+    kale.carpet(samples[ii], ax=ax, color=h2[0].get_facecolor(), shift=ax.get_ylim()[0])
+
[…]
