+++ docs/source/api.rst
@@ -18,27 +18,7 @@ Each of these functions constructs a `KDE` (kalepy.kde.KDE) instance, calls the 
     new_samples = kde.resample()      # use same `KDE` for resampling
 
 
-.. include:: plot_api.rst
-
-
-Density estimation: `kalepy.density(...)`:
-------------------------------------------
-
-.. autofunction:: kalepy.density
-    :noindex:
-
-Data resampling: `kalepy.resample(...)`:
-----------------------------------------
-
-.. autofunction:: kalepy.resample
-    :noindex:
-    
-The KDE class: `kalepy.KDE`
----------------------------
-
-.. autoclass:: kalepy.KDE
-    :members:
-    :noindex:
+.. include:: kde_api.rst
 
 
 
