+++ docs/source/kalepy_kde.rst
@@ -1,6 +1,30 @@
 kalepy.kde
 ==========
 
+.. contents:: :local:
+
+Density estimation: `kalepy.density(...)`:
+------------------------------------------
+
+.. autofunction:: kalepy.density
+    :noindex:
+
+Data resampling: `kalepy.resample(...)`:
+----------------------------------------
+
+.. autofunction:: kalepy.resample
+    :noindex:
+    
+The KDE class: `kalepy.KDE`
+---------------------------
+
+.. autoclass:: kalepy.KDE
+    :members:
+    :noindex:
+
+Full submodule documentation
+----------------------------
+
 .. automodule:: kalepy.kde
    :members:
    :undoc-members:
