+++ _README_append.md
@@ -1,0 +1,18 @@
+
+## Development & Contributions
+
+Please visit the `github page <https://github.com/lzkelley/kalepy>`_ for issues or bug reports.  Contributions and feedback are very welcome.
+
+
+## Attribution
+
+A JOSS paper has been submitted.  If you have found this package useful in your research, please add a reference to the code paper:
+
+.. code-block:: tex
+
+    @article{kalepy,
+      author = {Luke Zoltan Kelley},
+      title = {kalepy: a python package for kernel density estimation and sampling},
+      journal = {The Journal of Open Source Software},
+      publisher = {The Open Journal},
+    }
