+++ docs/paper/paper.md
@@ -37,12 +37,12 @@ where the kernel can typically be expressed as,
 $$K_H(x) = \|H\|^{-1/2} K\left(H^{-1/2} x \right).$$
 Here $H$ is the 'bandwidth' (or covariance) matrix.  Choosing the kernel and bandwidth matrix produces most of the nuance and art of KDE.  The most common choice of kernel is likely the Gaussian, i.e.,
 $$\hat{f}_H(x) = \sum_{i=1}^N \frac{w_i}{(2\pi)^{-d/2}\|H\|^{1/2}} \exp\{(x_j - x_{ij}) {H^j}_k (x^k - {x_i}^k)\}.$$
-In the current implementation, the Guassian, tri-weight, and box-car kernels are implemented, in addition to the Epanechnikov kernel [@Epanechnikov-1969] which in some cases has been shown to be statistically optimal but has discontinuous derivatives that can produce both numerical and aesthetic problems.
+In the current implementation, the Gaussian, tri-weight, and box-car kernels are implemented, in addition to the Epanechnikov kernel [@Epanechnikov-1969] which in some cases has been shown to be statistically optimal but has discontinuous derivatives that can produce both numerical and aesthetic problems.
 Often the bandwidth is chosen to be diagonal, and different rules-of-thumb are typically used to approximate a bandwidth that minimizes typical measures of error and/or bias.  For example, the so-called 'Silverman factor' [@Silverman-1978] bandwidth,
 $$H_{ij} = \delta_{ij} \sigma_i \left[ \frac{4}{(d + 2)n}\right]^{1/(d+4)} \;\;\;\textrm{(summation not implied)},$$
 where $\delta_{ij}$ is the Kronecker delta, and $\sigma_i$ is the standard deviation (or its estimate) for the $i$th parameter.  In the current implementation, both the Silverman and Scott factor [@Scott-1979] bandwidth estimators are included.
 
-Reflecting boundary conditions can be used to improve reconstruction accuracy.  For example, with data drawn from a log-normal distribution, a standard KDE will produce 'leakage' outside of the domain.  To enforce the restriction that $f(x < 0) = 0$ (which must be known {\textit{a priori}}), the kernel is redefined such that $K_H(x < 0) = 0$, and re-normalized to preserve unitarity\footnote{Note that some implementations instead truncate and renormalize the resulting $\hat{f}_H$ which which incorrectly redistributes probability from near the boundaries to the whole domain.}.  This example is shown in Fig.\autoref{fig:one}, with histograms in the upper panel and KDEs on the bottom.
+Reflecting boundary conditions can be used to improve reconstruction accuracy.  For example, with data drawn from a log-normal distribution, a standard KDE will produce 'leakage' outside of the domain.  To enforce the restriction that $f(x < 0) = 0$ (which must be known {\textit{a priori}}), the kernel is redefined such that $K_H(x < 0) = 0$, and re-normalized to preserve unitarity\footnote{Note that some implementations instead truncate and renormalize the resulting $\hat{f}_H$ which which incorrectly redistributes probability from near the boundaries to the whole domain.}.  This example is shown in \autoref{fig:one}, with histograms in the upper panel and KDEs on the bottom.
 
 Resampling from the derived PDF can be done much more efficiently in the KDE framework than by the standard method of CDF inversion.  In particular, we can see that sampling from the PDF is identical to re-sampling with replacement from the weighted data points, while shifting each point based on the PDF of the Kernel at that location.
 
