+++ README.md
@@ -101,7 +101,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_8_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_8_0.png)
 
 
 ### resampling: constructing statistically similar values
@@ -126,7 +126,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_11_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_11_0.png)
 
 
 ### Multivariate Distributions
@@ -146,7 +146,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_13_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_13_0.png)
 
 
 
@@ -175,7 +175,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_14_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_14_0.png)
 
 
 # Fancy Usage
@@ -210,7 +210,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_18_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_18_0.png)
 
 
 Explicit reflection locations can also be provided (in any number of dimensions).
@@ -242,7 +242,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_20_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_20_0.png)
 
 
 ### Multivariate Reflection
@@ -264,7 +264,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_22_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_22_0.png)
 
 
 ### Specifying Bandwidths and Kernel Functions
@@ -312,7 +312,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_24_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_24_0.png)
 
 
 ## Resampling
@@ -357,7 +357,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_27_0.png)
+![png](https://raw.githubusercontent.com/lzkelley/kalepy/dev/docs/media/demo_files/demo_27_0.png)
 
 
 ### Resampling while 'keeping' certain parameters/dimensions
@@ -379,7 +379,7 @@ nbshow()
 ```
 
 
-![png](https://raw.githubusercontent.com/lzkelley/kalepy/test/docs/media/demo_files/demo_29_0.png)
[…]
