+++ docs/paper/paper.bib
@@ -282,3 +282,15 @@ archivePrefix = {arXiv},
   title = {corner.py: Scatterplot matrices in Python},
   journal = {Journal of Open Source Software}
 }
+
+@article{GetDist-2019,
+ author         = "Lewis, Antony",
+ title          = "{GetDist: a Python package for analysing Monte Carlo
+                   samples}",
+ year           = "2019",
+ eprint         = "1910.13970",
+ archivePrefix  = "arXiv",
+ primaryClass   = "astro-ph.IM",
+ SLACcitation   = "%%CITATION = ARXIV:1910.13970;%%",
+ url            = "https://getdist.readthedocs.io"
+}
