+++ docs/paper/paper.md
@@ -25,9 +25,10 @@ bibliography: paper.bib
 
 # Statement of need
 
-Numerous python KDE implementations exist, for example in `scipy` (`scipy.stats.gaussian_kde`) [@SciPy-2020], `seaborn` (`seaborn.kdeplot`) [@seaborn-2020], and most notably `KDEpy` [@kdepy-2018] which provides excellent performance on large numbers of data points and dimensions.  None of these packages, however, provide built-in resampling or boundary conditions, and no single package provides both plotting and statistical results.  `kalepy` is entirely class-based, and while focusing on ease of use, provides a highly extensible framework for modification.
+Numerous python KDE implementations exist, for example in `scipy` (`scipy.stats.gaussian_kde`) [@SciPy-2020], `seaborn` (`seaborn.kdeplot`) [@seaborn-2020], and most notably `KDEpy` [@kdepy-2018] which provides excellent performance on large numbers of data points and dimensions.  None of these packages, however, provide built-in resampling or boundary conditions, and no single package provides both plotting and statistical results.  The `GetDist` package offers extensive methods for plotting samples, utilizing KDEs with numerous boundary treatments [@GetDist-2019], but without resampling or a standalone KDE interface. `kalepy` is entirely class-based, and while focusing on ease of use, provides a highly extensible framework for modification and expansion in a range of possible applications.
 
 While `kalepy` has no features specific to any particular field, it was designed for resampling from weighted astronomical datasets.  Consider a population of binaries derived from cosmological simulations.  If the initial population is costly to produce (e.g. requiring tens of millions of CPU hours), and as long as it accurately samples the parameter space of interest, it may be sufficiently accurate to produce larger populations by 'resampling with variation', e.g. using a KDE approach.  Depending on the details of the population, many of the parameters may be highly correlated and often abut a boundary: for example, the mass-ratio defined as the lower-mass component divided by the more massive component, is often highly correlated with the total mass of the binary, and is bounded to the unit interval i.e. $q \equiv M_2 / M_1 \leq 1$.  Faithfully resampling from the population requires handling this discontinuity, while also preserving accurate covariances which may be distorted when transforming the variable, performing the KDE, and transforming back.
+
 
 # Methods
 
