### 2020-12-30 1d1b84f
Fix deprecation from newer pandoc
