+++ docs/paper/build.sh
@@ -11,4 +11,4 @@ ENGINE="xelatex"
 OPTS="-V geometry:margin=1in"
 
 # pandoc --filter pandoc-citeproc --bibliography=paper.bib --pdf-engine=xelatex -s paper.md -o paper.pdf
-pandoc --filter pandoc-citeproc ${OPTS} --bibliography=${INPUT_BIB} --pdf-engine=${ENGINE} -s ${INPUT_TXT} -o ${OUTPUT_PDF}
+pandoc --citeproc ${OPTS} --bibliography=${INPUT_BIB} --pdf-engine=${ENGINE} -s ${INPUT_TXT} -o ${OUTPUT_PDF}
