### 2020-12-30 86a7989
DOCS: add github link in sphinx index
