### 2020-12-30 73d84c3
DOCS/BUG: fix issue with image filenames in index.rst
