### 2020-12-30 fb4f47a
Update 'CHANGES.md' file
