### 2020-12-30 #2784 @lzkelley
Dear @kexinrong and @williamjameshandley, thank you both so much for your helpful comments!  I have just pushed a new version 'v0.5' to master which includes the recommended updates.  The changes are noted in the [CHANGES.md file](https://github.com/lzkelley/kalepy/blob/master/CHANGES.md#v05---20201230), and also here:

- [kalepy.readthedocs.io](kalepy.readthedocs.io) now includes fully compiled documentation from docstrings, in addition to usage instructions and numerous examples.  These are primarily included on the [API](https://kalepy.readthedocs.io/en/latest/api.html) page, which now also includes examples on resampling (and its features) and plotting.
- The API to the plotting submodule has been completely restructured for a far-simpler API.  The plotting functionality is also highlighted more prominently (and clearly) in the [readthedocs plotting examples](https://kalepy.readthedocs.io/en/latest/api.html#plotting-distributions).
- I have fixed the typos in the JOSS paper, and discussed the `GetDist` package in particular.  I believe the most important difference is that `GetDist` is a **plotting package** that uses KDEs (very nicely), while `kalepy` is a **KDE package** that includes plotting routines.  In addition to the added functionality of resampling, I wasn't able to find any user-interface to the KDE functionality in `GetDist` *per se*, outside of plotting routines.
- In both the README and readthedocs index, I have included information about both contributions and attributions.  For the time being, I have not added any codes of conduct, bug-report templates, etc as I don't expect these to be necessary for the scale of community contributions that I expect on this relatively-small package.
