### 2021-01-01 #2784 @williamjameshandley
This is very nearly there in my book. At the moment, the readthedocs website doesn't seem to be extracting the docstrings. For an example of what I mean, have a look [here](https://anesthetic.readthedocs.io/en/latest/anesthetic.html), and how this is automated [here](https://github.com/williamjameshandley/anesthetic). There still aren't any clear community guidelines for contributions (other than a small subsection). I recommend following the advice [here](https://github.com/lzkelley/kalepy/community).
