### 2021-01-02 8ec1d87
Fix issue with sphinx module docs
