+++ kalepy/kernels.py
@@ -1,4 +1,28 @@
-"""Kernal basis functions for KDE calculations.
+"""Kernal basis functions for KDE calculations, used by `kalepy.kde.KDE` class.
+
+Contents:
+
+- :class:`Kernel <kalepy.kernels.Kernel>` :
+  class performing the numerical/mathematical functions of a KDE using a particular kernel-function.
+
+- :class:`Distribution <kalepy.kernels.Distribution>` :
+  base class for kernel-function functionality
+
+- :class:`Gaussian(Distribution) <kalepy.kernels.Gaussian>` :
+  class for Gaussian kernel functions
+
+- :class:`Box_Asym(Distribution) <kalepy.kernels.Box_Asym>` :
+  class for box (top-hat) kernel functions
+
+- :class:`Parabola(Distribution) <kalepy.kernels.Parabola>` :
+  class for parabolic (Epanechnikov) kernel functions
+
+- :func:`get_distribution_class <kalepy.kernels.get_distribution_class>` :
+  returns the appropriate `Distribution` subclass matching the given string specification.
+
+- :func:`get_all_distribution_classes <kalepy.kernels.get_all_distribution_classes>` :
+  returns a list of active `Distribution` subclasses (used for testing).
+
 """
 import logging
 import six
@@ -13,6 +37,12 @@ from kalepy import _NUM_PAD, _TRUNCATE_INFINITE_KERNELS
 
 
 _INTERP_NUM_PER_STD = int(1e4)
+
+
+__all__ = [
+    'Kernel', 'Distribution', 'Gaussian', 'Box_Asym', 'Parabola',
+    'get_distribution_class', 'get_all_distribution_classes'
+]
 
 
 class Kernel(object):
