+++ kalepy/plot.py
@@ -13,19 +13,35 @@ fairly large number of keyword arguments for customization.  The top level API m
 additional customization is possible by using the drawing methods directly, and optionally
 subclassing the `Corner` class to provide additional or different functionality.
 
-API Contents
+Plotting API
 ------------
-- `Corner` : class for corner/triangle/pair plots.
-- `corner` : method which constructs a `Corner` instance and plots 1D and 2D distributions.
 
-- `dist1d` : plot a 1D distribution with numerous possible elements (e.g. histogram, carpet, etc)
-- `dist2d` : plot a 2D distribution with numerous possible elements (e.g. histogram, contours, etc)
+- :class:`Corner <kalepy.plot.Corner>` :
+  class for corner/triangle/pair plots.
 
-- `carpet` : draw a 1D scatter-like plot to semi-quantitatively depict a distribution.
-- `contour` : draw a 2D contour plot. A wrapper of additional functionality around `plt.contour`
-- `confidence` : draw 1D confidence intervals using shaded bands.
-- `hist1d` : draw a 1D histogram
-- `hist2d` : draw a 2D histogram.  A wrapper of additional functionality around `plt.pcolormesh`
+- :func:`corner <kalepy.plot.corner>` :
+  method which constructs a `Corner` instance and plots 1D and 2D distributions.
+
+- :func:`dist1d <kalepy.plot.dist1d>` :
+  plot a 1D distribution with numerous possible elements (e.g. histogram, carpet, etc)
+
+- :func:`dist2d <kalepy.plot.dist2d>` :
+  plot a 2D distribution with numerous possible elements (e.g. histogram, contours, etc)
+
+- :func:`carpet <kalepy.plot.carpet>` :
+  draw a 1D scatter-like plot to semi-quantitatively depict a distribution.
+
+- :func:`contour <kalepy.plot.contour>` :
+  draw a 2D contour plot. A wrapper of additional functionality around `plt.contour`
+
+- :func:`confidence <kalepy.plot.confidence>` :
+  draw 1D confidence intervals using shaded bands.
+
+- :func:`hist1d <kalepy.plot.hist1d>` :
+  draw a 1D histogram
+
+- :func:`hist2d <kalepy.plot.hist2d>` :
+  draw a 2D histogram.  A wrapper of additional functionality around `plt.pcolormesh`
 
 """
 
@@ -1593,8 +1609,8 @@ def draw_hist2d(ax, edges, hist, mask_below=None, **kwargs):
     return edges, hist, rv
 
 
-def draw_contour2d(ax, edges, hist, quantiles=None, smooth=None, upsample=None, pad=True,
-                   outline=True, **kwargs):
+def draw_contour2d(ax, edges, hist, quantiles=None, levels=None,
+                   smooth=None, upsample=None, pad=True, outline=True, cbar=None, **kwargs):
 
     LW = 1.5
 
@@ -1631,7 +1647,9 @@ def draw_contour2d(ax, edges, hist, quantiles=None, smooth=None, upsample=None, 
     edges = [xx, yy]
 
     # ---- Setup parameters
-    _, levels, quantiles = _dfm_levels(hist, quantiles=quantiles)
+    if levels is None:
+        _, levels, quantiles = _dfm_levels(hist, quantiles=quantiles)
+
     alpha = kwargs.setdefault('alpha', 0.8)
     lw = kwargs.pop('linewidths', kwargs.pop('lw', LW))
     kwargs.setdefault('linestyles', kwargs.pop('ls', '-'))
[…]
