+++ docs/source/kalepy.rst
@@ -1,40 +1,44 @@
-=========================
-full package documenation
-=========================
+kalepy full package documentation
+=================================
 
+.. contents:: :local:
 
-Primary API
------------
+kalepy.kde module
+---------------------
 
-`See the API documentation here. <api>`_
+.. automodule:: kalepy.kde
+    :members:
+    :undoc-members:
+    :show-inheritance:
 
+kalepy.kernels module
+----------------------
+
+.. automodule:: kalepy.kernels
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+kalepy.plot module
+-------------------------
+
+.. automodule:: kalepy.plot
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+kalepy.utils module
+-----------------------
+
+.. automodule:: kalepy.utils
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. Module contents
+.. ---------------
+.. 
 .. .. automodule:: kalepy
-..    :members:
-..    :imported-members:
-
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Submodules:
-
-   kalepy_kde
-   kalepy_kernels
-   kalepy_plot
-   kalepy_utils
-
-   
-
-.. Submodules
-.. ==========
-.. 
-.. `kalepy.kde <kalepy_kde.html>`_
-.. -------------------------------
-.. 
-.. `kalepy.kernels <kalepy_kernels.html>`_
-.. ---------------------------------------
-.. 
-.. `kalepy.plot <kalepy_plot.html>`_
-.. ---------------------------------
-.. 
-.. `kalepy.utils <kalepy_utils.html>`_
-.. -----------------------------------
+..     :members:
+..     :undoc-members:
+..     :show-inheritance:
