### 2021-01-02 6c54688
Expand contributions/development information
