### 2021-01-02 502b6f1
BUG: trying to fix 'No module named 'kalepy'' error on readthedocs
