### 2021-01-02 #2784 @lzkelley
Apologies @williamjameshandley, there was a readthedocs build error that I didn't notice.  The docs seem to be building properly now.  The [full docstring documentation is here](https://kalepy.readthedocs.io/en/latest/kalepy.html).  I have expanded the contribution information in the [README](https://github.com/lzkelley/kalepy#development--contributions), the [DEVELOPMENT.rst](https://github.com/lzkelley/kalepy/blob/master/DEVELOPMENT.rst) file, and in the [readthedocs index](https://kalepy.readthedocs.io/en/latest/#development-contributions).
