### 2021-01-02 a7bced7
BUG: should not invert colors when adding colorbar contours from draw_contour2d
