### 2021-01-03 #2784 @kexinrong
Hi Luke, thanks for the changes. The updated documentation and guideline look good to me.

For the JOSS paper, I have one suggestion regarding discussions on related packages: it seems that `scipy.stats.gaussian_kde` does have a build-in [resample](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.gaussian_kde.resample.html) function. Could you update the writing to reflect this?
