### 2021-01-07 2ecbb1c
BUG: Correct description of other KDE packages
