### 2021-01-07 #2784 @lzkelley
@kexinrong thanks for pointing that out!  I clarified (and cleaned up) the discussion about other packages.  New version is pushed to master.
