### 2021-01-16 #2784 @arfon
@lzkelley - I made a collection of small fixes to the paper in https://github.com/lzkelley/kalepy/pull/1
