### 2021-01-18 d7d2b6b
Add 'outline' kwarg to 'dist2d()' to allow path_effects to be disabled
