### 2021-01-18 f412e3a
Add 'outline' kwarg to 'dist2d()' to allow path_effects to be disabled
