### 2021-01-19 d406cfb
BUG: fix memory error in '_parse_extrema', cleanup handling of 'reflect'
