+++ kalepy/kernels.py
@@ -115,12 +115,13 @@ class Kernel(object):
                 np.shpae(weights), num_data)
             raise ValueError(err)
 
+        # `reflect` should be sanitized and converted from calling method, not here!
         # if (reflect is not None) and (len(reflect) != npar_data):
         #     err = "Length of `reflect` ({}) does not much data dimensions ({})!".format(
         #         len(reflect), npar_data)
         #     raise ValueError(err)
-        reflect = _check_reflect(reflect, data, weights=weights)
-
+        # reflect = _check_reflect(reflect, data, weights=weights)
+        
         # -----------------    Calculate Density
 
         whitening = sp.linalg.cholesky(matrix_inv)
@@ -147,6 +148,11 @@ class Kernel(object):
 
         # -------------------   Perform Reflection
 
+        if (len(reflect) != npar_data):
+            err = "ERROR: shape of reflect `{}` does not match data `{}`!".format(
+                np.shape(reflect), np.shape(data))
+            raise ValueError(err)
+        
         for ii, reflect_dim in enumerate(reflect):
             if reflect_dim is None:
                 continue
