### 2021-01-19 e5450d6
BUG: fix memory error in '_parse_extrema', cleanup handling of 'reflect'
