### 2021-01-21 050cfc5
Allow existing 'axes' to be passed to Corner initializer
Minor formatting fixes.
