### 2021-01-21 669263c
Allow 'reflect' to be shaped (2,) instead of (1,2) for 1D input data
