### 2021-01-21 25df4d9
Pass bandwidth from KDE to '_guess_edges' when possible, to use as part of bin-width estimate
