### 2021-01-21 dbceb6f
BUG: 'edges' was not being used in density plot in dist1d
