### 2021-01-22 fa2ed13
Minor paper tweaks
