+++ kalepy/VERSION.txt
@@ -1,1 +1,1 @@
-0.5
+1.0.0
