### 2021-01-22 #2784 @lzkelley
Thanks for the corrections @arfon, I've created the release v1.0.0 and associated [zenodo DOI: 10.5281/zenodo.4456130
](https://zenodo.org/record/4456130).  I fixed another typo, and couple of grammatical issues in the draft.
